Extract stream_buffer pure helpers into a shared module
Move buffer_body / compute_stream_delta (and their private helpers) out of
orchestrator.py into a new stream_buffer.py so both the CLI orchestrator and the
upcoming Agent SDK driver share one implementation. Relocate the pure-function
unit tests into stream_buffer_test.py; the CLI-specific _StreamBufferConsumer
test stays in orchestrator_test.py. No behavior change.
